Refactored to avoid use of std::function.
I found out that std::function can create overhead,
even including unwanted heap allocations. This is not
a good idea for a realtime audio renderer.
I found out from a conversation with Microsoft Copilot
a way to rework the code to eliminate this problem,
although it does require a little extra typing.